Fix warnings and errors during compile
I fixed all of the errors and warnings that came up while trying to
compile. I added the extern keyword to board.h definitions of variables
that were defined in board.c. I also fixed a couple of warnings that
showed up due to (x = x & (x-1)) having unsafe parentheses being used.
The change turned it into (x &= (x - 1)). There was one other example of
this, again with an assignment in an if statement.