Add tests for boardMove
I added a bunch of tests for boardMove, and in doing so discovered some
bugs that I took the time to fix. I also added some new macros to aid
us, such as a counter part to the square names: A1 now has IA1 which is
the index of that square. This is useful because Move uses an index to
specify src and dst. Additionally I discovered that we use two
definitons of BLACK and WHITE: 1 and 0, as well as 6 and 0. This led to
some headaches as I was trying to get the color from a move to then
index into the pieces array in a Board. I changed the mgetcol to return
6 instead of 1 for black, as I anticipate we will use that more often. I
  also added _WHITE and _BLACK for the bitwise color variants.