En pessant now generated correctly
When we changed en pessant, we added a flag for when en pessant was
possible rather than relying on the en pessant square. This flag wasn't
always checked, especially in perftRun so en pessant was either generated
when there was no en pessant in getPieceAttackMap or counted when there
wasn't en pessant in perftRun. printBoardInfo now prints "None" when
there is no en pessant.