Perft on position 2 depth 3 is now correct
To accomplish this, I compared the positions that lefoux generated with
the positions leape generated for discrepencies. To output a position,
there is a printFen function and a commandline command for it "fen".
The command line also has the uci "position" command implemented the
same way that leape has it implemented. Move now has the mgettaken and
mgetprevinfo macros for better readability.

The issue was with detecting en passant which we previously did by
simply checking if the square is the same as the en passant square,
however the en passant square defaults to A3 so an extra move was
generated attacking A3 when en passant wasn't possible there. To fix
this, macro bgetenp now returns a 1 when there is en passant and a 0
when there isn't. After making this change, it was just a matter of
tracking down where en passant is being checked.